# openapi.yaml
openapi: 3.0.0
info:
  title: Feedback 360 Survey API
  version: 2.0.0
  description: |
    This API specification contains information around the Feedback 360 Survey
    API developed by the App Dev Practice.

    The Feedback 360 Survey API enables the management of feedback surveys for
    customer engagements. A feedback survey is N/A to 5 point scale of competency around
    the technologies used for a project. Each added employee will be able to enhance the
    survey by adding technologies, softskills, and freetext for each of his or her
    colleagues. The will enable Red Hat to better understand multiple aspects of
    a consultant's skillset.
  contact:
    email: appdevpractice@redhat.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
servers:
  - url: "http://localhost:3007"
    description: LOCAL
  - url: "http://survey-service-consultant-360-dev.apps.shared-dev.dev.openshift.opentlc.com"
    description: Spring Boot Service that implements the Feedback 360 Survey API on DEV Cluster
  - url: "http://survey-service-consultant-360-prod.apps.shared-dev.dev.openshift.opentlc.com"
    description: Spring Boot Service that implements the Feedback 360 Survey API on PROD Cluster
paths:
...
  /employee:
    get:
      tags:
        - employee
      parameters:
        - name: filterBy
          description: Field by which to filter results.
          in: query
          schema:
            type: string
          example: name
        - name: filter
          description: String to filter on, query string
          in: query
          schema:
            type: string
          example: Frank
        - name: sortBy
          description: Field by which to sort
          in: query
          schema:
            type: string
        - name: sortOrder
          description: Sort Order
          in: query
          schema:
            type: string
            enum:
              - ASC
              - DESC
        - name: offset
          description: Page offset
          schema:
            type: integer
            format: int32
          in: query
        - name: maxResults
          description: Maximum number of results to return, defaults to 20
          schema:
            type: integer
            format: int32
          example: 20
          in: query
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
          description: A list resource collection of Employees.
      operationId: getEmployees
      summary: List all employees available in the Feedback 360 Survey API.
      description: >
        This operation fetches a full list of employees available in the Feedback 360 Survey API.  The
        filtering and sorting mechanism for fetching this list is yet To Be Determined.  The Employees
        in the Feedback 360 Survey API data store will be populated from outside resources such as
        LDAP or Rover.
  "/employee/{employeeId}":
    get:
      tags:
        - employee
      parameters:
        - name: employeeId
          description: Kerberos ID of the employee.
          schema:
            type: string
            example: ajones
          in: path
          required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
          description: A representation of the employee matching the given employeeId.
        "404":
          description: An employee matching the given employeeId was not found.
      operationId: getEmployeeById
      summary: Find an Employee resource by ID
      description: Returns a single employee matching the given ID.
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
      tags:
        - employee
      parameters:
        - name: employeeId
          description: Kerberos ID of the employee.
          schema:
            type: string
            example: ajones
          in: path
          required: true
      responses:
        "200":
          description: >-
            The employee represented by the supplied userid has been updated with the contents
            of this request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: >-
            The employee was not updated. Something is wrong with the request like a missing
            required field.
        "404":
          description: This employee was not found
      operationId: updateEmployeeByID
      summary: |
        Update an Employee resource by ID
      description: >
        The representation of the employee contained in this request will replace the employee
        located at this URI. This operation is idempotent.
...
  "/employee/{employeeId}/surveyGroups":
    get:
      tags:
        - employee
      parameters:
        - name: employeeId
          description: Employee's RedHat Kerberos ID.
          schema:
            type: string
            example: ajones
          in: path
          required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeAssignmentWithSurveyGroup"
          description: A List of Survey Groups matching the given employeeId.
        "404":
          description: An employee matching the given employeeId was not found.
      operationId: getSurveyGroupByEmployeeId
      summary: Find an all Survey Groups an employee is part of resource by ID
      description: Returns a List of Survey Groups matching the given employee ID.
...
components:
  schemas:
...
    Employee:
      type: object
      x-java-class-annotations:
        - "@javax.persistence.MappedSuperclass"
      x-tablename: employees
      properties:
        id:
          description: Employee's RedHat Kerberos ID
          type: string
          example: jsmith@redhat.com
          x-java-field-annotations:
            - "@javax.persistence.Id"
          x-primary-key: true
        name:
          description: Name of the employee.
          type: string
          example: John Smith
        email:
          description: The email address of the employee.
          type: string
          example: jsmith@redhat.com
          x-java-field-annotation:
            - |-
              "@org.hibernate.annotations.Formula("CONCAT(id, '@redhat.com')")"
        role:
          description: "The role of the employee. For example, consultant, PM, TSM, etc."
          type: string
          example: Consultant
        hasAcceptedTerms:
          description: Whether the employee has accepted the terms of the usage agreement.
          type: boolean
          example: false
          x-java-field-annotations:
            - "@javax.persistence.Transient"
        dateTermsAccepted:
          description: The date the employee accepted the terms of the usage agreement.
          type: string
          format: date
          example: "2021-02-09"
        termsVersionNumber:
          description: The version number of terms of the usage agreement.
          type: number
          format: float
          example: 1.0
    EmployeeAssignmentWithSurveyGroup:
      type: object
      allOf:
        - $ref: "#/components/schemas/SurveyGroup"
      properties:
        startDate:
          description: The start date of this employee's assignment
          type: string
          format: date-time
        endDate:
          description: The end date of this employee's assignment
          type: string
          format: date-time
        billableRole:
          type: string
          description: The billing role that is assigned for the project (May not match actual title)
          # enum:
          # - Associate Consultant
          # - Consultant
          # - Senior Consultant
          # - Architect
          # - Principal Consultant
          # - Senior Architect
          # - Principal Architect
          # - Distinguished Consultant
          # - Distinguished Architect
          # - Project Manager
...
    SurveyGroup:
      required:
        - opportunityId
        - projectName
        - projectCreatorId
        - tsmId
      type: object
      x-java-class-annotations:
        - "@javax.persistence.MappedSuperclass"
        - "@javax.persistence.EntityListeners(org.springframework.data.jpa.domain.support.AuditingEntityListener.class)"
      properties:
        id:
          description: A GUID that uniquely identifies the SurveyGroup.
          type: string
          format: uuid
          example: f1ad7649-eb70-4499-9c82-a63fe2c6dc71
          x-java-field-annotations:
            - "@javax.persistence.Id"
            - '@javax.persistence.GeneratedValue(generator = "UUID")'
            - '@org.hibernate.annotations.GenericGenerator(name = "UUID", strategy =
              "org.hibernate.id.UUIDGenerator")'
            - '@javax.persistence.Column(name = "id", updatable = false, nullable =
              false)'
        createdDate:
          description: The date the project was created in the Feedback 360 Survey API in UTC.
          type: string
          format: date-time
          readOnly: true
          example: "2020-01-29"
          x-java-field-annotations:
            - "@org.springframework.data.annotation.CreatedDate"
        createdBy:
          description: The employee id (Kerberos) of the user that created the project.
          type: string
          example: janedoe
          x-java-field-annotations:
            - "@org.springframework.data.annotation.CreatedBy"
        modifiedDate:
          description: The date the project was last modified in the Feedback 360 Survey API in UTC.
          type: string
          format: date-time
          readOnly: true
          example: "2020-01-29"
          x-java-field-annotations:
            - "@org.springframework.data.annotation.LastModifiedDate"
        modifiedBy:
          description: The employee id (Kerberos) of the user that last modifed the project.
          type: string
          example: janedoe
          x-java-field-annotations:
            - "@org.springframework.data.annotation.LastModifiedBy"
        opportunityId:
          description: The ID of the Opportunity from PSA.
          type: string
          example: 3456NAS
          x-java-field-annotations:
            - "@javax.persistence.Column(unique = true)"
        projectName:
          description: The name of the project.
          type: string
          example: NASA App Modernization
        projectCreatorId:
          description: The email address of the creator of the project.
          type: string
          example: janedoe@redhat.com
        tsmId:
          description: The email address of the TSM in charge of the project.
          type: string
          example: timmytsm@redhat.com
        defaultSkills:
          type: array
          x-java-field-annotations:
            - "@javax.persistence.ManyToMany"
            - |-
              @javax.persistence.JoinTable(
                  name = "default_survey_skills",
                  joinColumns = @javax.persistence.JoinColumn(name = "survey_group_id"),
                  inverseJoinColumns = @javax.persistence.JoinColumn(name = "skill_id"),
                  uniqueConstraints = @javax.persistence.UniqueConstraint(columnNames = {"survey_group_id", "skill_id"})
              )
          items:
            $ref: "#/components/schemas/Skill"
    Skill:
      type: object
      x-java-class-annotations:
        - "@javax.persistence.Entity"
        - '@javax.persistence.Table(name = "skills")'
      properties:
        id:
          description: A GUID that uniquely identifies a skill.
          type: string
          format: uuid
          example: d567521f-13a6-4237-936c-40bbbf388bc5
          x-java-field-annotations:
            - "@javax.persistence.Id"
            - '@javax.persistence.GeneratedValue(generator = "UUID")'
            - '@org.hibernate.annotations.GenericGenerator(name = "UUID",strategy = "org.hibernate.id.UUIDGenerator")'
            - '@javax.persistence.Column(name = "id", updatable = false, nullable = false)'
        skillsBaseId:
          description: An id from skills base
          type: integer
          example: 1234
        skill:
          description: The name of the skill.
          type: string
          example: Crucial Conversations
        description:
          description: A description of the skill.
          type: string
          example: The ability to have crucial conversations with clients.
          x-java-field-annotations:
            - "@javax.persistence.Column(length=512)"
        category:
          description: The general category of the skill.
          type: string
          example: leadership
        active:
          description: Boolean to state if this skill is active
          type: boolean
          example: true
          default: true
...